10711 moved files to more logical folder locations.
