+++ frontend-react/src/shared/LiveMap/LiveMap.module.scss
@@ -1,4 +1,4 @@
-@use "../../../global-modules.scss" as *;
+@use "src/global-modules" as *;
 
 .legend {
     list-style: none;
